25.Design a program using concepts of inter-process communication ordinary pipes in which one process sends a string message to a second process, and the second process reverses the case of each character in the message and sends it back to the first process. For example, if the first process sends the message Hi There, the second process will return hI tHERE. This will require using two pipes, one for sending the original message from the first to the second process and the other for sending the modified message from the second to the first process. You can write this program using either UNIX or Windows pipes
Pipes:-
Pipes is nothing but connection between two processors. pipe acts as a conduit allowing two processes to communicate. 
In the pipes it behaves first in frist out. In data structure pipe behave like queue. In pipes size of read and write don’t have to match in the pipe in the pipe we can write in 512 bytes at a time but we can read 1 byte at a time in a pipe.
In the operating system we can use a pipe such that one process is used to write the pipe. In the other process reads from the pipe.
It opens a pipe which is an area of main memory that is treated as a “virtual file”. The pipes is used by creating process, and its child processes for reading and writing one process can write to virtual file or pipe and another related process can read from it.
Pipe is nothing but a system call it finds the frist two available positions in the process’s one is open file table and allocates them for the read and write the pipes at the end.
This is about the pipes in operating system.
Advantages:
In pipes it is synchronized naturally.
It is use is simple and it is essy to create also.
In this pipes no extra system calls required to communicate (read or write).
Disadvantages:
It is having less memory size like 4k.
In pipes only related process can communicate.
In pipes it communicate with only two process and it communicate in one direction.
In pipes there is no involvement of kernel.